pyproject.toml
[tool.poetry]
name = "spacepaperwork"
version = "0.1.0"
description = ""
authors = ["Nigel Gott <nigel.gott@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.4"
django-allauth = "^0.44.0"
django-autocomplete-light = "^3.8.1"
django-dbbackup = "^3.3.0"
django-environ = "^0.4.5"
django-extensions = "^3.1.0"
django-money = "^1.2.1"
django-timezone-field = "^4.1.1"
psycopg2 = "^2.8.6"
pycodestyle = "^2.6.0"
django-fsm = "^2.7.1"
djangorestframework = "^3.12.2"
gsheets = "^0.5.1"
django-prometheus = "^2.1.0"
django-materializecss-form = "1.1.17"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
python-dateutil = "^2.8.1"
google_auth_oauthlib = "^0.4.2"
django-tinymce = "^3.2.0"
bravado = "^11.0.2"
django-contrib-comments = "^2.0.0"
django-tenants = "^3.2.1"


[tool.poetry.dev-dependencies]
Werkzeug = "^1.0.1"
mypy = "^0.790"
django-stubs = "^1.7.0"
pytest = "^6.2.1"
pytest-sugar = "^0.9.4"
selenium = "^3.141.0"
splinter = "^0.14.0"
pytest-pythonpath = "^0.7.3"
pytest-watch = "^4.2.0"
watchdog = "^1.0.2"
freezegun = "^1.0.0"
requests-mock = "^1.8.0"
sphinx = "^3.4.0"
sphinx-autobuild = "^2020.9.1"
flake8 = "^3.8.4"
black = "^20.8b1"
pre-commit = "^2.9.3"
prospector = "^1.3.1"
django-debug-toolbar = "^3.2"
pytest-django = "^4.1.0"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
